backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
requests
python-jose[cryptography]
bcrypt
